third_party/blink/web_tests/external/wpt/css/css-overflow/scroll-markers/scroll-marker-activation-retains-focus-expected.txt

This is a testharness.js-based test.
[FAIL] CSS Test: ::scroll-marker retains focus on ::scroll-marker
  assert_equals: ::scroll-marker activation retains focus expected "rgb(0, 0, 255)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
